# pkg/api/v1/rest/projects/api_projects.go
/*
 * OpenSource Issue Träcking System
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 * Contact: DevBoxFanBoy@github.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package projects

import (
	"encoding/json"
	"errors"
	"fmt"
	"github.com/DevBoxFanBoy/opists/pkg/api/router"
	"github.com/DevBoxFanBoy/opists/pkg/api/v1"
	"github.com/DevBoxFanBoy/opists/pkg/api/v1/model"
	"github.com/gorilla/mux"
	"net/http"
	"strings"
)

// A ApiController binds http requests to an api service and writes the service results to the http response
type ApiController struct {
	service v1.ProjectsApiServicer
}

// NewApiController creates a default api controller
func NewApiController(s v1.ProjectsApiServicer) router.Router {
	return &ApiController{service: s}
}

// Routes returns all of the api route for the ApiController
func (c *ApiController) Routes() router.Routes {
	return router.Routes{
		{
			"GetAllProject",
			strings.ToUpper("Get"),
			"/v1/projects",
			c.GetAllProject,
		},
		{
			"CreateProject",
			strings.ToUpper("Post"),
			"/v1/projects",
			c.CreateProject,
		},
		{
			"GetProject",
			strings.ToUpper("Get"),
			"/v1/projects/{projectKey}",
			c.GetProject,
		},
		{
			"UpdateProject",
			strings.ToUpper("Put"),
			"/v1/projects/{projectKey}",
			c.UpdateProject,
		}, {
			"DeleteProject",
			strings.ToUpper("Delete"),
			"/v1/projects/{projectKey}",
			c.DeleteProject,
		},
	}
}

// GetAllProject - Returns all projects
func (c *ApiController) GetAllProject(w http.ResponseWriter, r *http.Request) {
	result, err := c.service.GetAllProject()
	if err != nil {
		router.InternalError(w, err)
		return
	}
	router.EncodeJSONResponse(result, nil, w)
}

// GetProject - Returns the Project by key
func (c *ApiController) GetProject(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	projectKey := params["projectKey"]
	if errorResponse, err := validateProjectKey(projectKey); err != nil {
		router.BadRequestErrorResponse(w, errorResponse.(model.ErrorResponse))
		return
	}
	result, err := c.service.GetProject(projectKey)
	if err != nil {
		router.NotFoundError(w, err)
		return
	}
	router.EncodeJSONResponse(result, nil, w)
}

// UpdateProject - Updates the Project by key
func (c *ApiController) UpdateProject(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	projectKey := params["projectKey"]
	if errorResponse, err := validateProjectKey(projectKey); err != nil {
		router.BadRequestErrorResponse(w, errorResponse.(model.ErrorResponse))
		return
	}
	body := &model.Project{}
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		router.InternalError(w, err)
		return
	}
	result, err := c.service.UpdateProject(projectKey, *body)
	if err != nil {
		errRes := result.(model.ErrorResponse)
		router.HandleErrorResponses(w, errRes)
		return
	}
	status := http.StatusNoContent
	router.EncodeJSONResponse(``, &status, w)
}

// CreateProject -  Create a new Project
func (c *ApiController) CreateProject(w http.ResponseWriter, r *http.Request) {
	prj := &model.Project{}
	if err := json.NewDecoder(r.Body).Decode(&prj); err != nil {
		router.InternalError(w, err)
		return
	}
	if errorResponse, err := validateProjectKey(prj.Key); err != nil {
		router.BadRequestErrorResponse(w, errorResponse.(model.ErrorResponse))
		return
	}
	result, err := c.service.CreateProject(*prj)
	if err != nil {
		errRes := result.(model.ErrorResponse)
		router.HandleErrorResponses(w, errRes)
		return
	}
	status := http.StatusCreated
	w.Header().Set("Location", fmt.Sprintf("%v/%v", r.RequestURI, prj.Key))
	router.EncodeJSONResponse(result, &status, w)
}

func (c *ApiController) DeleteProject(w http.ResponseWriter, r *http.Request) {
	params := mux.Vars(r)
	projectKey := params["projectKey"]
	if errorResponse, err := validateProjectKey(projectKey); err != nil {
		router.BadRequestErrorResponse(w, errorResponse.(model.ErrorResponse))
		return
	}
	result, err := c.service.DeleteProject(projectKey)
	if err != nil {
		errRes := result.(model.ErrorResponse)
		router.HandleErrorResponses(w, errRes)
		return
	}
	status := http.StatusNoContent
	router.EncodeJSONResponse(``, &status, w)
}

func validateProjectKey(projectKey string) (interface{}, error) {
	if len(projectKey) == 0 {
		err := errors.New(fmt.Sprintf("ProjectKey %v is invalid!", projectKey))
		return model.ErrorResponse{
			Code:    400,
			Message: err.Error(),
		}, err
	}
	return nil, nil
}